requirements.txt
Pillow>=7.1.2
Flask>=1.1.2